[chore] #253 Rename canvas package name to tooltip package
